ALCN-68 ALCN-80 extract lot and stock
(cherry picked from commit c64bd2312106e1dc15bcd5eeca293a5bf2124527)

+++ import_db2/importer.py
@@ -91,7 +91,8 @@ class Importer:
             file_path = os.path.join(self.csv_path, '%s.csv' % name)
 
             # backup the file before overwriting it to make diff csv files
-            copyfile(file_path, file_path + '.former')
+            if os.path.isfile(file_path):
+                copyfile(file_path, file_path + '.former')
             with open(file_path, 'wb') as csvfile:
                 writer = csv.DictWriter(csvfile, headers, lineterminator='\n')
                 writer.writeheader()
